+++ src/components/datepicker/datepicker.js
@@ -850,6 +850,7 @@ DatePicker.prototype = {
 
         const cell = $(this);
         cell.addClass(`is-selected${(s.range.useRange ? ' range' : '')}`).attr('aria-selected', 'true');
+        self.lastValue = null;
         self.insertSelectedDate(cell);
 
         if (s.range.useRange) {
@@ -1656,6 +1657,7 @@ DatePicker.prototype = {
    */
   setToday(keepFocus) {
     const s = this.settings;
+    this.lastValue = null;
     this.currentDate = new Date();
 
     if (!this.settings.useCurrentTime) {